Feature/update mqtt (#131)
This completes the feature-set for updating, by adding updates over MQTT to both the asset hierarchy and raw. It adds a single "Update" config option to the bridge, which enables or disables updating assets or timeseries that are pushed. The update is conservative, only overwritting fields that are non-null in the new object, and merging metadata.

To make this work in a general way for raw-rows was trickier. In the end, it worked fairly well to treat both new and old as raw json objects, then merge them recursively. This works well for this case.

As usual there is some overhead with enabling updating, but that is expected, and probably not a problem since this is done fairly rarely.