Feature/type handling (#134)
Separate logic for creating data-types out of the main extractor class. Add support for enums and for auto-discovery of numeric types during runtime.

This might seem like overkill, but it is necessary for better type-support in the future, and centralizing the data-type logic is nice. BufferedDataTypes are now shared, so if multiple timeseries have the same type they all have pointers to the same object. While we don't really use this yet, it is technically more memory efficient, and it is in general a better design (less duplicated information).

Moving the logic means that we can add a few missing features:

 - Null datatype as numeric, this has been an issue for a customer once (though it absolutely is a bug in the server). It may be a little edge-case, but for some servers it might be useful.
 - DataType as metadata. Knowing the datatype can potentially be useful for grouping timeseries.

Also proper support for enums. Without an explicit mapping, most enums would be mapped to the string representation of their numeric value, which is never going to be correct. With AutoDiscovery enabled, they will now be mapped to either pure numeric values, or optionally to their label values.

If they are mapped to numbers, the label values are added as metadata. This highlighted an issue: the limiting of metadata length was wrong, this is also fixed.