Read ArrayDimensions for properties
This caused array properties to not get populated if MaxArraySize was
not set. The reason we need to read ArrayDimensions at all for
properties is that properties may be ByteString, which can represent
arbitrary files, and we don't really want to download gigabytes of files
if we can avoid it.