Allow configuring regex filters for converting variables to properties (#138)
Treat variables as properties with a regex filter on either name or string id. Some variables are variables not because they vary from the source system, but because they represent configurable options. These are sometimes more useful as properties. This is a general enough solution that it should work quite well.

By chance this also uncovered a bug. If array variables were disabled and some properties were arrays, the value of these would not be read no matter what, since ArrayDimensions were not read.

Co-authored-by: Einar Omang <einar@omang.com>
Co-authored-by: Einar Omang <einar.omang@cognite.com>