Fix really old, stubborn bug with timestamp
This has been the case for a long time and the solution explains why:
For some reason, when adding a % symbol in a pre build event, you have
to type it as %25. On windows, you actually have to do it twice, so to
represent %ai, to parse the author commit time, you have to do %25%25ai.
If you use only a single %25 you get a compile-time error.

On linux and OSX, on the other hand, you have to use %25ai, only once,
or you get %ai as just text instead of a meaningful output.

I found no good way to solve this in general, so I just added a
condition and duplicated the pre build event.