Refactor configuration of events
Event emitters and types are no longer added manually, but automatically
detected on startup using the EventNotifier property. This is the
correct way to do it, and much easier to use. The ConfigTool used this
method, but it also used other methods, which isn't really necessary.
The specification states that EventNotifier must be set.

The EventEmitters, EventTypes, HistorizingEventEmitters properties have
been removed, and replaced with Enabled, AllTypes (indicating whether to
extract all event types), a regex filter on DisplayName for ignoring
event types, and a History option to enable reading history from
historizing emitters.

Most users will be interested in extracting as many types as possible,
so it makes sense that the default is extracting everything, with the
exception of history which should be enabled manually, since it is
potentially very heavy.