Re-add removed config for backwards compatibility
It now works as a whitelist for events and additional features. Some
functionality is slightly changed, but for anyone using only the
generated event config (which is the case for everyone using the
extractor) it should be essentially the same.

The changes are that it is no longer possible to disable certain
emitters. This is not a big issue, as it is not possible to disable
certain historical nodes for datapoints.