Two part fix to an observed crash on rebrowse
First, repeated crashes resulted in an ugly loss of connection to the
server, which at some point caused the extractor to be unable to stop.
The old way of quitting the extractor was poor, the new implementation
uses a linked cancellationtokensource, which is much more idiomatic.

Second, the crash itself was caused by an objectreferenceexception (with
very little logging). I was able to at least get the same error with
similar circumstances, this is probably the cause. The issue was
comparing values of properties that had not been initialized due to too
large dimensions, or other limitations, when updating. I added a test
that reproduced the error, then fixed it.