Update dependencies (#150)
Involves a few changes due to updated linter, these are microsoft recommended linter settings, and the same as used in the utils, so it is nice to obey them and be consistent.

Adds support for token based authentication to CDF.

Some refactoring was necessary due to the large changes to the utils. Most of the changes are just in the .lock file though.