Add chunking to event subscriptions (#153)
This is necessary now since event subscriptions are detected automatically. Earlier it was unlikely that anyone would configure >1000 distinct event emitters, but now it _could_ happen. (it is still very unlikely, most servers only have _one_). This is an edge-case.

This also takes the opportunity to unify the code for creating data-point and event subscriptions. The refactor is most of the changes, and essentially just means pulling most of the code that was SubscribeToNodes into a separate method and making it slightly more generic, then calling that from both SubscribeToNodes and SubscribeToEvents.