Avoid updating to 5.0
This is a problem that will need to be fixed. The current build server
does not support 5.0, and I'm not sure we can in good conscience update
the extractor to 5.0, however it will not build with 5.0 packages unless
the version on the build server is updated, which means making a new
docker image.