Fix/misc bugfix (#159)
Fix a couple of issues with the extractor:
 - Clean up behavior if using the example config.. Ideally you should be able to use the example config without it affecting anything, but some of that is just not feasible. Another patch to verify config a little is needed.
 - Fix issue where datapoints would be lost if they were timestamped before 1970. This was a part of the extractor before, but disappeared at some point.
 - Update packages, resolves issue with null metadata. Also use the new chunkByHierarchy method from the utils. It is better written and quite a bit more efficient.
 - Make the config tool behave a bit better if the server does not support events. The config tool should be able to fail, with the consequence of just disabling that part, not crashing and failing to create a config file.

The actual changes are pretty small. The lines changed are almost all just the lock file.

This ended up being a lot more complicated as well, because the SDK now uses packages from .net 5.0, so running the tests required an update to .net core. Presumably the same issue may happen if you try to run the service, but the standalone should not be affected.

Previously the build used our own dotnet-mono image, but paket no longer requires mono, so we don't really need it, and can instead just use the dotnet image provided by microsoft, which is far more future-proof. Next step is updating the core packages to 5.0 as well.