Test rework for UAClient (#158)
First step in a refactor of the extractor tests. The issue is that the current setup means every single test is an integration test, which slows down development and means that quite a few tests are inconsistent. Instead, we should use a common server context where possible. This adds unit tests for the UAClient class, which can allow us to remove integration tests later once the other parts of the system also have proper unit tests.

This effort also exposed a few potential issues in UAClient.