Fix/reduce memory usage (#166)
When dealing with the memory leak on a very large server, the amount of memory used to represent the node hierarchy was large (~2GB). This isn't ideal, as most of this data isn't always useful. It was used for a few things:

 - Check for changes on re-browse - now handled using a checksum.
 - Push missing nodes when a pusher fails to initialize on startup - added a separate buffer to the pushers that store the nodes they need to push. This is more flexible anyway. It might need a little work to be perfect.
 - Get the parentId of variables in MQTTPusher, for events. - needed to store them separately.

I also changed things a little so we reuse the stored Id on the node-state for datapoints. This was a very simple change and could actually be quite significant.

This is also updated with similar changes for relationships.