Feature/type definitions (#170)
Feature requested by customer. Nodes in OPC-UA have a TypeDefinition, which in some cases may be useful to know about. This is a simple solution to adding them. In the future it may be relevant to also add them as assets, then create a relationship, but that is a more complex feature.

This simply adds a "TypeDefinition" metadata field to every node. It involved a little extra code to read the type definition of root nodes.

Most of the changes here was a necessary refactor: split the TypeCollectors file into a file per class, and put them in a separate folder and namespace. This makes the code a fair bit cleaner, and I'm expecting further additions to these in the future.