Extractor test rework (#172)
Refactor the UAExtractor class a little, and add tests for its public methods.

The main change is to remove any dependency on UAExtractor from UAClient. This was already mostly done, the final required change was to remove the callbacks on reconnect and disconnect. Doing this using events instead is a lot smoother and more elegant, and makes it easier to use UAClient for other things (such as future improvements to the ConfigTool).

Also modified the PushNodes method a little, so that only nodes which weren't actually pushed are saved. This is probably an unlikely edge-case, but it doesn't hurt, especially with a thorough test.