Core unit tests (#173)
This is the second part of the major test-refactoring effort. This PR updates and refactors based on changes required for the following core components of the extractor:

 - The Streamer, which handles subscription datapoints and events
 - The Looper, which handles periodic events
 - ExtractorRuntime, which is a convenience class containing some of the common initialization code for the service and the launcher. It also ensures that the extractor is properly disposed if we need to restart.
 - FailureBuffer, which handles datapoints and events triggered on non-historical nodes while the connection to CDF is down.
 - HistoryReader, which manages reading historical data from CDF.

Each of these have gotten unit tests for most of their methods, where possible, and in the process have been refactored for cleaner code and in some cases to fix bugs.

This PR is big, though most of it is just the tests. There is one commit for each component.