Pusher tests (#175)
Rework of tests on the pushers and PusherUtils.

The refactoring here is not that extensive. The biggest change is moving the conversion of nodes into CDF-objects to the node classes themselves, which is cleaner. Some superfluous code was also removed.

Most of this PR is just a whole lot of tests, on each of the pushers. Some small changes in order to make them more similar. The data stored in each destination should be similar. This is especially true for MQTT/CDF, since the destination there is the same.