Integration tests (#176)
Finally, remove the old tests and replace them in part with new ones. The general idea of this test rework has been that integration tests should be limited to performing large tests that verify general functionality. A part of this has been to separate the pushers from the integration tests completely. It is not ideal that we rely on one pusher to verify the functionality of the others.

The new integration tests use the DummyPusher to verify the pushed contents, and rely on unit tests on the individual pushers to verify that the extracted data is pushed correctly to destinations.

The only old tests that remain are the tests for the ConfigurationTool, which is going to be reworked completely, and the old tests for the influxdb buffer. The influxdb buffer works poorly, adds way too much complexity, and isn't really used anywhere, so it should really be retired.

I also ran the visual studio code cleanup function, which is a fair number of these changes.