More complex types (#179)
Make an effort to improve support for more complex return types. This particular issue was just triggered by Opc.Ua.KeyValuePair, which we now have explicit support for, but in general there should be a better way to handle more complex types. This uses the built-in system to convert complex types to json (it is used by the OPC-UA JSON protocol), but only if we can't do it on our own. The result isn't always proper JSON, and it can be a bit weird, but it is a lot better than just getting the type name in CDF.

There are a few tests to make sure this works as intended across the board. In general we now pass the raw DataValue.WrappedValue Variant to ConvertToString. This way we can get the TypeInfo and properly parse the result. The performance hit _should_ be minimal.