Source transformations (#181)
Flexible system for transforming the source node hierarchy.

Each customer appears to have their own desires and requirements for what they want to be treated as properties. The old system was limited by the fact that only some limited subset of node context was available in the HandleNode method.

This new system aims to remove limitations on how to ignore or treat nodes as properties. First, objects can now be properties, meaning that it is possible to arbitrarily nest properties. It is fairly common in OPC-UA servers representing control systems to have "folders" of metadata for an asset representing an asset, this allows us to treat the entire folder and any sub-folders as metadata.

Second, the old "ignoreName" and "ignoreNamePrefix" options have been deprecated in favor of a list of Regex-based filters. Which allows advanced users to do the same as before, but with more flexibility.

The old system still works, but now gives a warning about deprecation.