Clean up history read (#183)
It is now more explicitly correct. This was in response to a potential issue where history results were mixed up. I don't think this was an issue before, the server in question has had multiple issues of this sort, but the code was not very clear. This is much more in the spirit of the OPC-UA library.