Remove SourceName from BaseExcludeProperties (#184)
Also make it configurable. Turns out SourceName is not required to match SourceNode. LocalTime and ReceiveTime are still excluded, but the BaseExcludeProperties field has been made writeable, so that we can in theory solve issues that require these as well. I don't really want to document this yet (making it official...), we might want a better solution if it turns out that it is necessary.