Read object types as assets (#186)
While there is limited useful information in the object types in base OPC-UA, most of the more complex data models use the object types to model common properties of objects. Especially the ones based on the ISA-95 standard. This means that extracting information from at least part of the node type hierarchy is highly relevant.

This change is not very large, but it is a bit delicate. Before, the node class has been implicit in the class. Now it is explicit, and while types are in many ways similar to normal nodes, they are different in a few subtle ways.

For example, it does not make sense to treat variable types as timeseries, but they do have most of the same properties as variables, except for "Historizing". So we want them treated as variables, but mapped to assets.

The result works well enough for a proof of concept. If necessary we can expand this further later.

It is debatable whether it is interesting to map variables on types to timeseries. A simple solution to this is to just add NodeClass as a filter option to source transformations, so that it is possible to ignore or transform them into properties easily.

Due to some issues when testing this, this also fixes an issue where assets updates to CDF would not be correctly sanitized.