Restructure node attributes (#188)
The goal of this change is to decouple the attribute set a bit more from the UANode and UAVariable classes, making it so that it is possible to refer to the same attribute set from several different nodes.

The abstraction helps in two ways:

First, it moves more of the internal node logic out of UAClient, by handling the attribute-read on the attributes class.

Second, it makes it so that array children can refer directly to the same attribute set as their parent node, which also means that they get the same properties, removing the need for the complex syncing of array parents and children.

The change itself is pretty small, but it unfortunately made the tests a bit more tricky, since a very large number of tests create local UANode or UAVariable objects. Almost all the changes are just moving things around a bit in the tests.