Deep properties (#189)
Support for reading deep hierarchies of properties under timeseries as well. This appears to be fairly common, and there is nothing in the standard that prevents it. Properties generally should not have properties of their own, so we skip browsing those for efficiency.

The new system is actually simpler than the old in some ways. We just call the BrowseHierarchy method on the given list of nodes and their properties, then add them as properties to their parents. BrowseHierarchy is already recursive, so the result is that we recursively map out properties.

For testing, the server is expanded with a deep metadata hierarchy under the NumberVar2 node.