Add support for reading history without continationPoints (#190)
This was requested. The implementation is okay, in the end, this isn't really something we want to recommend, as it is not a compliant way to read history from the servers. That said, it is not unlikely that some servers have problems with reading history normally, and so this is a decent solution.

Read history by simply adjusting the start time parameter to the most conservative value given the list of states being extracted after each iteration. It has some weird behavior when reading multiple nodes at the same time, but it does work. It also requires the time for events to be correct, so it would not really work for the use case this was originally requested for. That said, less advanced servers may easily have trouble with this, so it makes sense to implement it.