Add support for nested event properties
Also fixes a very serious bug in the event implementation. Turns out the
previous approach was completely wrong, and a re-read of the relevant
parts of the standard reveals that the best way to handle event select
clauses for us is to just specify BaseEventType as TypeId always, which
is going to work just fine. We don't allow excluding event properties
based on type, which is what would have been the use-case of using the
event id. This would make the SelectClause much larger, which would
probably reduce performance.

The actual nested event implementation was not very complicated, but
EventFields are now unique by BrowsePath rather than just BrowseName.