Add support for nested event properties (#191)
Also fixes a very serious bug in the event implementation. Turns out the previous approach was completely wrong, and a re-read of the relevant parts of the standard reveals that the best way to handle event select clauses for us is to just specify BaseEventType as TypeId always, which is going to work just fine. 

Essentially, the issue was that we were selecting event properties based on original type/browse path pairs, while the correct approach is selecting based on subtype/browse path pairs. Alternatively, it is possible to just specify the BaseEventType as type, which works as a kind of wildcard.

We don't allow excluding event properties based on type, which is what would have been the use-case of using the subtypeid. This would make the SelectClause much larger, which would probably reduce performance.

The actual nested event implementation was not very complicated, EventFields are now unique by BrowsePath rather than just BrowseName.