Certificate auth (#197)
Certificate based authentication. This is kind of a half-baked solution. It works, and it is testable, but in the end it turns out that this is very highly server specific, so any implementation will be based to some extent on guesswork. The OPC-UA SDK has a system to sign authentication requests using x509 certificates, and this implements that. Other forms of authentication will have to wait.