History throttling (#198)
System for throttling history reads using three different parameters:

 - Max number of nodes in parallel. This is independent of chunking, but a useful metric since throttling is mainly about limiting the load on the server.
 - Max number of parallel requests.
 - Max number of initiated requests per minute.

The last two are implemented using the utils TaskThrottler.

Throttling is per type of history read, so MaxParallelism = 2 means potentially 4 requests in parallel if the extractor reads both data-points and events at the same time. This was mainly for simplicity, but I think it is actually desirable in general. Dividing the numbers by two is perfectly fine, and this way the reading will be evenly distributed. Chances are also pretty good that event history load and datapoint history load may be different on the server, so this utilizes the resources on the server better.

The idea is also that this system can be extended to other operations in the future (subscriptions, Read, and Browse).

This PR also by necessity makes some changes required by a new version of the utils.