Serialize to json (#202)
Concept for serializing nodes fully to json. Before, nested properties were serialized as:

`prop1_prop2_prop3: "nested value"`

but while this is well suited for extracting directly to clean, it doesn't work that well when extracting to Raw. Instead, this should be converted to a nested JSON object, which can be manipulated by transformations and the like.

The OPC-UA SDK contains some logic for transforming complex data types to JSON, which is used. The complexity comes from having to preserve the old logic as well, without duplicating too much code. The ConvertToString method from UAClient is moved to a new class, UAPropertySerializer, which behaves slightly differently depending on whether it serializes to JSON or string.

For now, metadata is JSON in Raw, and the old form in Clean, due to limitations on metadata value length.