Add option to write NodeIds to Raw
This is useful for a couple of purposes. By writing the full NodeIds to
raw we make it possible to read them back later, for buffering. It can
also be helpful to retrieve the full NodeIds for configuration purposes.

Add a few tests and move old StringConverter tests from UAClientTest to
StringConversionTest.