Refactor json serialize (#208)
Refactor JSON serialization to raw to use a JsonConverter class. The better way to do this would be to use System.Text.Json's JsonConverter, but that won't be possible until .NET 6, since we need the WriteRawValue method.

For now this just removes a feature: the ability to meaningfully filter which fields should get updated in Raw. This feature had little real purpose anyway. In general this simplifies the code quite a bit, avoiding the use of AssetCreate and StatelessTimeseriesCreate when serializing to raw.

The main goal of doing this is increasing the flexibility of the data we send to Raw, to make it possible to have use-cases that do not fit neatly into simply extending AssetCreate. The ability to flexibly change the data we push to Raw opens new possibilities.