Test improvements (#211)
Some fixes to make tests easier to write, and remove some old code.

I added a method to reset the config in-place, without having to recreate the UAClient. This required some fancy reflection magic.

Also upgraded the old LiteDBStorageTests, which were the only tests still using the old tester, so it could be removed.

A few utility methods were moved to the BaseExtractorFixture class.

The old config.events.yml file was removed.