Multiplatform service (#209)
Use new .net 5 BackgroundService to integrate the service in the ExtractorLauncher, and make it available as a systemd service on Linux.

The installer and service manager applications have also been adapted for this. Some modifications were required to the installer:
 - It no longer starts the service automatically on install, since it requires a config.yml file to start at all now.
 - The folder structure has been modified slightly.

This neatly reduces the complexity of the code while adding some interesting new features.

The old embedded version system was also changed (it never worked as intended), and now the necessary information is properly embedded as AssemblyInformationalVersion and AssemblyDescription.