Startup from cdf (#216)
Get timeseries and event emitters from CDF on startup, instead of the OPC-UA node hierarchy.

The PR is pretty big, though at least 60% is in Test/. This is a feature that touches on the core of the extractor, and it requires some fairly thorough tests. To minimize duplicated code, BrowseResultHandler and the new CDFNodeSource now inherit from a BaseNodeSource class with some common attributes, this is referenced from UAExtractor when browsing starts.

Nodes get a Source attribute, which refers to where they originated. This is needed so that nodes from CDF are not pushed back to CDF (without metadata, causing a potential loss of information).

In theory you could configure the extractor to read from CDF Raw and push to CDF Clean, though we should consider that an unintended and unsupported feature.