Improve memory usage (#228)
A customer noticed high memory usage peaks during rebrowse. This is likely due to reading from Raw before updating. Now, rows are read and consumed in chunks, which allows the GC to reclaim more memory earlier, if needed.

This PR also includes some server code used to test higher memory usage, as well as a fix to build the MSI version of the extractor as 64 bit.