Static analysis (#222)
This is part of the prep for open sourcing the extractor. Enabling some additional static analysis helps with guaranteeing code quality. This PR is mostly Nullable analysis, meaning that nullability of reference types must be explicitly stated, with some additional available compiler hints.

This helps restrict the code, and generally encourages good practices. The analysis is a bit limited, however, so it is disabled or worked around in some specific places. The solution to these would require very large changes to the code.

The main issue is the configurability. Very many fields can be null due to missing config. Solving this properly would require a lot of additional abstraction that may not really help with code readability.