System to wait for config file to exist (#230)
This is useful because the service tends to be started through the service manager, and it is ugly to have it dying when no valid config is found. This isn't really an ideal solution, but it is the old behavior before the rewrite of the service, and it is what users tend to expect.

This also adds a few checks to verify that the config is valid. More can be added here later.