Nodeset source (#229)
The OPC-UA standard defines a series of formats for storing nodes as xml files. The most general one is NodeSet2, which can be used to describe node hierarchies in detail. Some servers support exporting the node hierarchy in this form, or use a standard type library which may be exported this way. It makes sense to enable the extractor to use these files as starting point for exploring the OPC-UA server.

This is generally intended for small, well defined servers. Not reading the type hierarchy (which is often just the base hierarchy), reduces load on the server. If the server can even export the primary instance hierarchy, then the load can be reduced further.

With this the startup of the extractor can be configured to only read some limited data from the server:

 - The server node itself
 - The values of any properties or variable types. (as these will typically not be set by the server)

And of course datapoints/events once the mapping is completed.