Throttle browse (#227)
Total rewrite of browse logic. The idea is that instead of sequentially browsing nodes, we now schedule a pool of browse operations, similarly to how it is done with history. Like history, the pool grows over time, and browses are scheduled until none are running or waiting.

This makes the browse operations a little more loose and a little less predictable, but in general it should improve performance, and also allow for scheduling parallel requests in a good way.