Event type name (#232)
The requested feature was just adding event type name to metadata, but this was a good opportunity to align the event types with the other "type" classes, where instances reference common, global types. The change here was relatively simple, just make the hashset of event fields a property of a UAEventType type, then reference that from the event. This may come in handy later as well.

As an aside, the memory usage should be largely unaffected. If there are lots of events in memory, then this might actually reduce usage a little since we no longer keep copies of the event type nodeid (NodeId objects quickly add up). When no events are in memory we might use a little bit more memory just because we are now storing the name of each event type, this is insignificant, however.

The actual change is just adding that field to the event metadata when writing to CDF.