Estimate array size (#233)
While we prefer using ArrayDimensions to determine array sizes. Many do not support this. There are two other methods we could use to determine a suitable array size:

 - The MaxArraySize property, which is different from ArrayDimensions in that it is not the fixed size, but the max size.
 - The actual length of the current value.

The former is much better than the latter, so we prefer that one.

This PR implements an option to read both for nodes with ValueRank suggesting array values, but no ArrayDimensions.