Handle missing nodes (#235)
Handle nodes that are reported during browse but do not exist. This is done in two stages, because of how the state is available at different times.

If BadNodeIdUnknown is reported during browse, it is ignored. This is fine, we don't really care that much about non-critical issues during browse.

If BadNodeIdUnknown is reported during Read, we set the node to "Ignored", meaning it will not get read any further or pushed to CDF.