Update browser (#240)
Update BrowseScheduler to use the SharedResourceScheduler from the utils. It is mostly straight forward. The new scheduler does the same as the old one except for the generic behavior. It now also uses an external shared resource, which means that if we were to browse in parallel we could still respect the continuation point limit.

Since the old scheduler was synchronous and the old one is not this involved a decent amount of async-propagation to other methods. This is mostly a good thing, since shared sync and async isn't ideal.

Since the generic scheduler asks for it, this PR also creates a way to abort browse operations cleanly, which was lacking before.