Create server configuration for pubsub
Means adding a new node hierarchy containing the configuration to the
Server node, as well as a system to report changes directly to the
pubsub manager.

This works, and is capable of publishing messages as both JSON and UADP
(UA Datagram Protocol, a binary format).

The next step would be reading this configuration from a client, and use
it to build a subscriber capable of reading from pubsub. The
configuration can (probably) be written back to a file and saved.
Alternative approaches to configuration when the server is inaccessible
can be explored, but this approach is canonical (described in the
standard), and only requires the OPC-UA server to expose its
configuration.

Of course, this requires us to have access to the server, which may
remove some of the point of pubsub to begin with. The configuration
schema appears to be fairly standard to OPC-UA, so it might be possible
to export it from some applications.

The goal of this project is to have a proof of concept, and at least be
able to tell potential customers that we support OPC-UA pubsub, even if
there is a colossal asterisk.