Migrate looper (#244)
Use the PeriodicScheduler from the utils instead of the old Looper class. This has been required in the OPC-UA extractor for a while, the ability to have tasks that can be triggered manually or on a timer, and a central place to monitor running background tasks. The implementation can get pretty complex, and it makes sense to move it to the Utils.

The Looper class now contains just the actual method running periodically, and wraps an instance of the scheduler.