Migrate history (#241)
Replace the old HistoryScheduler with one based on the SharedResourceScheduler from the utils.

It works slightly differently, but should generally be fairly similar. An advantage of this is that the MaxNodeParallelism option now no longer applies separately to events and datapoints, which is more in line with the resource this actually represents (history continuation points, which are shared).

The PR is slightly bloated by the fact that this also moves history related classes into a separate history folder and namespace.