Create server configuration for pubsub (#242)
Means adding a new node hierarchy containing the configuration to the Server node, as well as a system to report changes directly to the pubsub manager.

This works, and is capable of publishing messages as both JSON and UADP (UA Datagram Protocol, a binary format) over MQTT. Using a different client and reading the json output, this can be confirmed to work manually.

The next step would be reading this configuration from a client, and use it to build a subscriber capable of reading from pubsub. The configuration can (probably) be written back to a file and saved. Alternative approaches to configuration when the server is inaccessible can be explored, but this approach is canonical (described in the standard), and only requires the OPC-UA server to expose its configuration.

Of course, this requires us to have access to the server, which may remove some of the point of pubsub to begin with. The configuration schema appears to be fairly standard to OPC-UA, so it might be possible to export it from some applications, but in general it is very complicated, due to its extreme level of generality.

This is a very fresh SDK feature, so it does not support alarms/events yet. I imagine this will be a requirement for actually using this properly later on, but it will probably be added to the SDK eventually.

The goal of this project is to have a proof of concept, and at least be able to tell potential customers that we support OPC-UA pubsub, even if there is a colossal asterisk.

This PR makes no changes to the extractor, except for a tiny fix related to an updated package. All the code touched is exclusively for testing.