Max history read length (#255)
This is helpful on servers with inefficient timeseries backends.

The implementation is a little tricky, but does not actually involve that much code. Essentially, the start time for each node is either equal to the furthest we've read, or the previous end time if the last read yielded no results.