Improve handling of failed browse a little (#268)
The issue occurred when connection was lost during browse. This still does not entirely resolve that, as the problem was that it took some time for the connection loss to be detected in the extractor class, this still happens, and there is no great way to avoid that that doesn't have potential side effects.

This PR improves logging a little, and handles this particular case better.