Remove late init properties (#271)
They are now read exclusively as part of the full browse, and not later when pushing. This is more in line with most use-cases and the performance cost when they are not is not that large. In practice this makes the code a lot simpler, and the logging much cleaner.

We had ended up reading them earlier in part or in full a lot more either way, the edge case had become the times we didn't need to read them on browse.