Rewrite attribute read (#277)
Rewrite system for reading attributes. The previous system was bug-prone and hard to understand. In fact, it had multiple bugs. These are fixed in this PR. Notably:

 - Variables with EventNotifier would use AccessLevel also for registering subscriptions to events.
 - Extractor would fail if history was enabled without data. (really bad)

Added a big test to verify that this still works, and that these issues are solved. The new system is also immune to the second type of error, where the extractor would fail due to a mismatch between GetAttributeIds and HandleAttributeRead.

Split ShouldSubscribe into ShouldSubscribeData and ShouldSubscribeEvents, the two are different enough in many ways that this is necessary.