Improve service logging (#278)
When reading logs, users will often see repeated "Reading attributes for [N] nodes..." or "Browse node hierarchy for ...", which is quite unclear, and makes it hard to understand what exactly the extractor is doing. Usually it logs some message before starting, but not always, and it is not always that helpful if multiple things are happening.

This adds a "Purpose" parameter to most of these service calls, so that when we log details we also log what operation is causing them. This helps a great deal.