Catch errors on WriteRawValue
This happened at a customer, but they were not able to provide any
information why, and they did something that removed the error. This
means there is an issue in ConvertToString somewhere, but it might be
very hard to find. To ensure it does not happen again, we should catch
and log errors on json serialization.