Recreate subscriptions that stop responding (#282)
This handles some unpleasant edge-cases on some servers. In theory this is illegal behavior, but it is also very unpleasant for customers, so we need some way to handle it. Ideally we would not do something that causes issues with servers that are working properly.

This also fixes an annoying bug in the trace log handler.